+++ watertap/examples/flowsheets/case_studies/electricity_costs.csv
